dev: + added time validation in serundeng + renamed getuk package to gaplek + more description in the READMES
